ch(rename): rename server/routes/users.js -> server/routes/signup.js
