+++ server/server.js
@@ -1,7 +1,8 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import morgan from 'morgan';
-import signUpRoutes from './routes/users';
+import signUpRoutes from './routes/signup';
+import signInRoutes from './routes/signin';
 import loanRoutes from './routes/loans';
 
 const app = express();
@@ -13,6 +14,7 @@ app.use(bodyParser.urlencoded({ extended: false }));
 
 // Routes which should handle requests
 app.use('/api/v1/signup', signUpRoutes);
+app.use('/api/v1/signin', signInRoutes);
 app.use('/api/v1/loans', loanRoutes);
 
 const PORT = process.env.PORT || 3000;
